# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Homer API
  version: 0.3.0
  description: Platform-agnostic multi-agent text service.
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
              example:
                ok: true
  /api/capabilities:
    get:
      summary: Runtime capabilities
      operationId: getCapabilities
      responses:
        "200":
          description: Active runtime capabilities and fallback visibility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
  /api/connectors/google_docs/auth/start:
    get:
      summary: Start Google Docs OAuth authorization flow
      operationId: startGoogleDocsOAuth
      responses:
        "200":
          description: OAuth flow initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorAuthStartResponse"
        "503":
          description: Google Docs OAuth is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                oauthUnavailable:
                  value:
                    error:
                      code: connector_service_unavailable
                      message: google docs oauth is not configured
  /api/connectors/google_docs/auth/callback:
    get:
      summary: Complete Google Docs OAuth authorization flow
      operationId: completeGoogleDocsOAuth
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: OAuth state returned by Google after consent.
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: OAuth authorization code returned by Google after consent.
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: OAuth provider error.
        - name: error_description
          in: query
          required: false
          schema:
            type: string
          description: Human-readable OAuth provider error message.
      responses:
        "200":
          description: OAuth callback succeeded and token was stored for the session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorAuthCallbackResponse"
        "400":
          description: Invalid callback payload or rejected consent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                missingState:
                  value:
                    error:
                      code: missing_oauth_state
                      message: state is required
                invalidState:
                  value:
                    error:
                      code: invalid_oauth_state
                      message: oauth state is invalid or expired
                accessDenied:
                  value:
                    error:
                      code: oauth_access_denied
                      message: access_denied
        "502":
          description: Authorization code exchange failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                exchangeFailed:
                  value:
                    error:
                      code: oauth_exchange_failed
                      message: oauth code exchange failed
        "503":
          description: Google Docs OAuth is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                oauthUnavailable:
                  value:
                    error:
                      code: connector_service_unavailable
                      message: google docs oauth is not configured
  /api/connectors/import:
    post:
      summary: Import document from configured connector
      operationId: importConnectorDocument
      description: Requires `X-Connector-Key` header when `CONNECTOR_API_KEY` is configured.
      security:
        - ConnectorApiKey: []
      parameters:
        - $ref: "#/components/parameters/ConnectorSessionHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorImportRequest"
      responses:
        "200":
          description: Connector document import succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorImportResponse"
        "400":
          description: Invalid payload or connector unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUnavailable:
                  value:
                    error:
                      code: connector_unavailable
                      message: connector credentials are unavailable
        "403":
          description: Connector access is forbidden for the target document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorForbidden:
                  value:
                    error:
                      code: connector_forbidden
                      message: connector access is forbidden for this document
        "404":
          description: Connector target document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorDocumentNotFound:
                  value:
                    error:
                      code: connector_document_not_found
                      message: connector document was not found
        "401":
          description: Connector request is not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUnauthorized:
                  value:
                    error:
                      code: connector_unauthorized
                      message: connector API key is invalid
        "502":
          description: Connector upstream credentials are invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUpstreamUnauthorized:
                  value:
                    error:
                      code: connector_upstream_unauthorized
                      message: connector upstream credentials are invalid
        "429":
          description: Connector request rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorRateLimited:
                  value:
                    error:
                      code: connector_rate_limited
                      message: connector rate limit exceeded
        "503":
          description: Connector upstream service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorServiceUnavailable:
                  value:
                    error:
                      code: connector_service_unavailable
                      message: connector service is unavailable
        "501":
          description: Connector import not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
  /api/connectors/export:
    post:
      summary: Export content through configured connector
      operationId: exportConnectorContent
      description: Requires `X-Connector-Key` header when `CONNECTOR_API_KEY` is configured.
      security:
        - ConnectorApiKey: []
      parameters:
        - $ref: "#/components/parameters/ConnectorSessionHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorExportRequest"
      responses:
        "200":
          description: Connector export succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorExportResponse"
        "400":
          description: Invalid payload or connector unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUnavailable:
                  value:
                    error:
                      code: connector_unavailable
                      message: connector credentials are unavailable
        "403":
          description: Connector access is forbidden for the target document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorForbidden:
                  value:
                    error:
                      code: connector_forbidden
                      message: connector access is forbidden for this document
        "404":
          description: Connector target document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorDocumentNotFound:
                  value:
                    error:
                      code: connector_document_not_found
                      message: connector document was not found
        "401":
          description: Connector request is not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUnauthorized:
                  value:
                    error:
                      code: connector_unauthorized
                      message: connector API key is invalid
        "502":
          description: Connector upstream credentials are invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorUpstreamUnauthorized:
                  value:
                    error:
                      code: connector_upstream_unauthorized
                      message: connector upstream credentials are invalid
        "429":
          description: Connector request rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorRateLimited:
                  value:
                    error:
                      code: connector_rate_limited
                      message: connector rate limit exceeded
        "503":
          description: Connector upstream service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                connectorServiceUnavailable:
                  value:
                    error:
                      code: connector_service_unavailable
                      message: connector service is unavailable
        "501":
          description: Connector export not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
  /api/task:
    post:
      summary: Run summarize or rewrite task
      operationId: runTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
            examples:
              summarize:
                summary: Summarize documents
                value:
                  task: summarize
                  documents:
                    - id: doc-1
                      title: Meeting notes
                      content: Q1 plan includes partner launch and hiring.
                  style: paragraph
                  instructions: Focus on action items.
                  enableCritic: false
              rewrite:
                summary: Rewrite text
                value:
                  task: rewrite
                  documents: []
                  text: We will utilize the platform to improve productivity.
                  mode: simplify
                  instructions: Keep it concise.
                  enableCritic: true
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: Invalid task payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
              examples:
                missingText:
                  value:
                    error:
                      code: missing_text
                      message: text is required for rewrite
                      requestId: 4e11fe43-e81c-40e8-b5cf-f9d4f0a65fe6
        "500":
          description: Internal processing failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorResponse"
components:
  securitySchemes:
    ConnectorApiKey:
      type: apiKey
      in: header
      name: X-Connector-Key
  parameters:
    ConnectorSessionHeader:
      name: X-Connector-Session
      in: header
      required: false
      description: Session key returned by Google Docs OAuth start/callback endpoints.
      schema:
        type: string
  schemas:
    Document:
      type: object
      required:
        - id
        - title
        - content
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
    TaskRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          enum:
            - summarize
            - rewrite
        documents:
          type: array
          description: Required for summarize tasks.
          items:
            $ref: "#/components/schemas/Document"
        text:
          type: string
          description: Required for rewrite tasks.
        mode:
          type: string
          description: Rewrite mode.
        instructions:
          type: string
        style:
          type: string
        enableCritic:
          type: boolean
          default: false
    PlanStep:
      type: object
      required:
        - id
        - role
        - action
      properties:
        id:
          type: string
        role:
          type: string
          enum:
            - planner
            - executor
            - critic
        action:
          type: string
    Metadata:
      type: object
      required:
        - provider
        - executionTimeMs
      properties:
        provider:
          type: string
          enum:
            - mock
            - openai
            - gemini
        executionTimeMs:
          type: integer
          format: int64
        requestId:
          type: string
    TaskResponse:
      type: object
      required:
        - result
        - plan
        - metadata
      properties:
        result:
          type: string
        plan:
          type: array
          items:
            $ref: "#/components/schemas/PlanStep"
        metadata:
          $ref: "#/components/schemas/Metadata"
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
    APIErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    CapabilitiesResponse:
      type: object
      required:
        - runtime
        - features
      properties:
        runtime:
          $ref: "#/components/schemas/RuntimeCapabilities"
        features:
          $ref: "#/components/schemas/FeatureFlags"
    RuntimeCapabilities:
      type: object
      required:
        - requestedProvider
        - activeProvider
        - providerFallback
        - requestedConnector
        - activeConnector
        - connectorFallback
      properties:
        requestedProvider:
          type: string
        activeProvider:
          type: string
          enum:
            - mock
            - openai
            - gemini
        providerFallback:
          type: boolean
        requestedConnector:
          type: string
        activeConnector:
          type: string
          enum:
            - none
            - google_docs
        connectorFallback:
          type: boolean
    FeatureFlags:
      type: object
      required:
        - critic
        - connectorImport
        - connectorExport
      properties:
        critic:
          type: boolean
        connectorImport:
          type: boolean
        connectorExport:
          type: boolean
    ConnectorImportRequest:
      type: object
      required:
        - documentId
      properties:
        documentId:
          type: string
    ConnectorAuthStartResponse:
      type: object
      required:
        - connector
        - sessionKey
        - authUrl
        - stateExpiresAt
      properties:
        connector:
          type: string
          enum:
            - google_docs
        sessionKey:
          type: string
        authUrl:
          type: string
          format: uri
        stateExpiresAt:
          type: string
          format: date-time
    ConnectorAuthCallbackResponse:
      type: object
      required:
        - connector
        - sessionKey
        - authenticated
      properties:
        connector:
          type: string
          enum:
            - google_docs
        sessionKey:
          type: string
        authenticated:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    ConnectorImportResponse:
      type: object
      required:
        - connector
        - document
      properties:
        connector:
          type: string
          enum:
            - google_docs
        document:
          $ref: "#/components/schemas/Document"
    ConnectorExportRequest:
      type: object
      required:
        - documentId
        - content
      properties:
        documentId:
          type: string
        content:
          type: string
    ConnectorExportResponse:
      type: object
      required:
        - connector
        - exported
      properties:
        connector:
          type: string
          enum:
            - google_docs
        exported:
          type: boolean